+++ src/dotnet/Rider.Plugins.CodeGenJetbrains/Execution/TestsGeneration/ComponentsIdentity.cs
@@ -1,0 +1,17 @@
+﻿namespace Rider.Plugins.CodeGenJetbrains.Execution.TestsGeneration;
+
+public static class ComponentsIdentity
+{
+    // searchable class pickers
+    public const string RequestActType = nameof(RequestActType);
+    public const string ResponseActType = nameof(ResponseActType);
+    public const string ResponseAssertType = nameof(ResponseAssertType);
+
+    // forms input fields
+    public const string ActEndpoint = nameof(ActEndpoint);
+    public const string AssertEndpoint = nameof(AssertEndpoint);
+
+    // misc. text boxes
+    public const string FilesPrefix = nameof(FilesPrefix);
+    public const string TestCaseName = nameof(TestCaseName);
+}