+++ src/dotnet/Rider.Plugins.CodeGenJetbrains/FluidModels/TestsGeneration/Models/Update/FUpdateTest.cs
@@ -1,0 +1,9 @@
+using Rider.Plugins.CodeGenJetbrains.FluidModels.TestsGeneration.Abstract;
+
+namespace Rider.Plugins.CodeGenJetbrains.FluidModels.TestsGeneration.Models.Update;
+
+public class FUpdateTest : FTest
+{
+    public FTestRequest Request { get; set; }
+    public FTestAssertRequest? AssertRequest { get; set; }
+}