+++ src/dotnet/Rider.Plugins.CodeGenJetbrains/FluidModels/TestsGeneration/Models/Update/FUpdateTestSuite.cs
@@ -1,0 +1,5 @@
+using Rider.Plugins.CodeGenJetbrains.FluidModels.TestsGeneration.Abstract;
+
+namespace Rider.Plugins.CodeGenJetbrains.FluidModels.TestsGeneration.Models.Update;
+
+public class FUpdateTestSuite : FTestSuite;