+++ src/rider/main/kotlin/com/jetbrains/rider/plugins/codegenjetbrains/actions/testsgeneration/GetListTestAction.kt
@@ -1,4 +1,4 @@
-﻿package com.jetbrains.rider.plugins.codegenjetbrains.actions
+﻿package com.jetbrains.rider.plugins.codegenjetbrains.actions.testsgeneration
 
 import com.jetbrains.rider.actions.base.RiderAnAction