+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 1.0.0
-- Initial version
+## 1.2.0
+- Added [GetList] test generation variant
+- Added [Update] test generation variant
+- Minor fixes and improvements
 
 ## 1.1.0
 - Added [GetList] test generation variant
 - Minor fixes and improvements
+
+## 1.0.0
+- Initial version